backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-tenants==3.6.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
django-celery-beat==2.6.0
boto3==1.34.144
django-storages==1.14.4
Pillow==10.4.0
reportlab==4.2.2
python-dotenv==1.0.1
dj-database-url==2.2.0
gunicorn==22.0.0
django-filter==24.2
requests==2.32.3
minio==7.2.7
